Move FileSystem implemenation to LoalFileSystem.
This change moves FileSystem's default implementation to
LocalFileSystem. As the result, some non-default FS lacks proper
implementation and fails. That is going to be fixed in coming
changes.

This is part of #2257.
